backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
psycopg2-binary==2.9.9
qdrant-client==1.7.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.2
spacy==3.7.2
tika==2.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
pandas==2.1.4
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
